Add dry run support for removing files

+++ python_scripts/filter_by_usr_name/cleanup.py
@@ -1,17 +1,25 @@
 import os 
 import glob 
+import argparse
+
+parser = argparse.ArgumentParser(description="Delete *_lines.txt files")
+parser.add_argument("--dry-run", action="store_true", help="Show files to be deleted without actually deleting them")
+args = parser.parse_args()
 
 pattern = "*_lines.txt"
 #glob.glob("*_lines.txt") find all the files with a specified name in current dir
 files = glob.glob(pattern)
 
 if not files:
-    print("There is no files to delete")
+    print("No matching files found")
 else:
     for file in files:
-        try:
-            os.remove(file)
-            print(f"Removed {file}")
-        except Exception as e:
-            print(f"Error occured while removing {file}: {e}")
+        if args.dry_run:
+            print(f"[DRY-RUN] Would remove: {file}")
+        else:
+            try:
+                os.remove(file)
+                print(f"Removed {file}")
+            except Exception as e:
+                print(f"Error removing {file}: {e}")
